Rename service result
