+++ blottertrax/value_objects/service_result.py
